app/requirements.txt
fastapi==0.115.12
httpx==0.28.1
joblib==1.5.0
numpy==2.2.5
pandas==2.2.3
pydantic==2.11.4
pydantic_core==2.33.2
pytest==8.3.5
pytest-cov==6.2.1
pytest-mock==3.14.0
python-multipart==0.0.20
scikit-learn==1.6.1
uvicorn==0.34.2
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-instrumentation-fastapi==0.54b1
opentelemetry-instrumentation-logging==0.54b1
opentelemetry-sdk==1.33.1
lightgbm==4.6.0